avoid x11 on bbb
