+++ pru/control/control.go
@@ -24,116 +24,29 @@ package main
 
 import (
 	"context"
-	"encoding/binary"
 	"flag"
 	"fmt"
 	"image"
 	"log"
 	"math/rand"
-	"os"
-	"sync/atomic"
-	"syscall"
 	"time"
-	"unsafe"
 
-	"fyne.io/fyne"
-	"fyne.io/fyne/app"
-	"fyne.io/fyne/canvas"
 	"github.com/fogleman/gg"
 	"github.com/jmacd/launchmidi/launchctl/xl"
 	"github.com/jmacd/nerve/pru/gpixio"
 )
 
-const deviceName = "/dev/rpmsg_pru30"
-
 var (
 	noInput = flag.Bool("no_input", false, "do not use MIDI input device")
-	noBBB   = flag.Bool("no_bbb", false, "do not use BeagleBone device")
 )
 
 type (
-	RPMsgDevice struct {
-		file *os.File
-	}
-
-	controlStruct struct {
-		framebufsAddr uint32
-		framebufsSize uint32
-		frameCount    uint32
-		dmaWait       uint32
-	}
-
 	Frameset    = gpixio.Frameset
 	Schedule    = gpixio.Schedule
 	Frame       = gpixio.Frame
 	DoubleRow   = gpixio.DoubleRow
 	DoublePixel = gpixio.DoublePixel
 )
-
-func openRPMsgDevice() (*RPMsgDevice, error) {
-	file, err := os.OpenFile(deviceName, os.O_RDWR, 0666)
-	return &RPMsgDevice{
-		file: file,
-	}, err
-}
-
-func (r *RPMsgDevice) write(data []byte) error {
-	n, err := r.file.Write(data)
-	if err != nil {
-		return nil
-	}
-	if n != len(data) {
-		return fmt.Errorf("short write: %d != %d", n, len(data))
-	}
-	return nil
-}
-
-func mmap(file *os.File, addr uint32, size int) ([]byte, error) {
-	return syscall.Mmap(
-		int(file.Fd()),
-		int64(addr),
-		size,
-		syscall.PROT_READ|syscall.PROT_WRITE,
-		syscall.MAP_SHARED,
-	)
-}
-
-func (r *RPMsgDevice) readControl() (*controlStruct, *Frameset, error) {
-	var data [32]byte
-	n, err := r.file.Read(data[:])
-	if err != nil {
-		return nil, nil, err
-	}
-	if n != 4 {
-		return nil, nil, fmt.Errorf("expected 4 bytes control address")
-	}
-	addr := binary.LittleEndian.Uint32(data[0:4])
-
-	mem, err := os.OpenFile("/dev/mem", os.O_RDWR, 0)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	cdata, err := mmap(mem, addr, int(unsafe.Sizeof(controlStruct{})))
-	if err != nil {
-		return nil, nil, err
-	}
-
-	ctrl := (*controlStruct)(unsafe.Pointer(&cdata[0]))
-
-	fdata, err := mmap(mem, ctrl.framebufsAddr, int(ctrl.framebufsSize))
-	if err != nil {
-		return nil, nil, err
-	}
-	if ctrl.framebufsSize != uint32(unsafe.Sizeof(Frameset{})) {
-		return nil, nil, fmt.Errorf("frameset size mismatch: C=%d, Go=%d", ctrl.framebufsSize, unsafe.Sizeof(Frameset{}))
-	}
-
-	framebuf := (*Frameset)(unsafe.Pointer(&fdata[0]))
-
-	// Note: should drop privileges now.
-	return ctrl, framebuf, nil
-}
 
 func Main() error {
 	var input *xl.LaunchControl
@@ -143,64 +56,15 @@ func Main() error {
 	if !*noInput {
 		input, err := xl.Open()
 		if err != nil {
-			log.Fatalf("error while opening connection to launchctl: %v", err)
+			return fmt.Errorf("error while opening connection to launchctl: %w", err)
 		}
 		defer input.Close()
 	}
 
-	var outputWindow fyne.Window
-	var inputWindow fyne.Window
-	var frames *Frameset
-	var inputImage *canvas.Image
-	var outputImage *canvas.Image
-	var buf = gpixio.NewBuffer()
-	outputPixels := image.NewRGBA(image.Rect(0, 0, 128, 128))
-
-	if *noBBB {
-		// frames is not in shared memory
-		frames = &Frameset{}
-
-		app := app.New()
-
-		inputWindow = app.NewWindow("Image")
-		inputImage = canvas.NewImageFromImage(buf.RGBA)
-		inputImage.FillMode = canvas.ImageFillOriginal
-		inputWindow.SetContent(inputImage)
-
-		outputWindow = app.NewWindow("Visage")
-		outputImage = canvas.NewImageFromImage(outputPixels)
-		outputImage.FillMode = canvas.ImageFillOriginal
-		outputWindow.SetContent(outputImage)
-
-	} else {
-		rpm, err := openRPMsgDevice()
-		if err != nil {
-			return err
-		}
-
-		// pru.c does not parse the message, this delivers the
-		// interrupt which causes pru.c to respond with its two
-		// carveout addresses.
-		if err := rpm.write([]byte("wakeup")); err != nil {
-			return err
-		}
-
-		var ctrl *controlStruct
-
-		ctrl, frames, err = rpm.readControl()
-		if err != nil {
-			return err
-		}
-
-		go func() {
-			before := atomic.LoadUint32(&ctrl.frameCount)
-			for {
-				time.Sleep(10 * time.Second)
-				after := atomic.LoadUint32(&ctrl.frameCount)
-				log.Println("frames/sec", after-before)
-				before = after
-			}
-		}()
+	buf := gpixio.NewBuffer()
+	state, err := newAppState(buf)
+	if err != nil {
+		return err
 	}
 
 	r := 0.8
@@ -261,25 +125,15 @@ func Main() error {
 			ggctx.SetRGB(r, g, b)
 			ggctx.Fill()
 
-			canvas.Refresh(inputImage)
+			buf.Copy(&state.frames[s])
+
+			state.test(s)
 
 			time.Sleep(33 * time.Millisecond)
-
-			buf.Copy(&frames[s])
-
-			testRender(&frames[s], outputPixels)
-
-			canvas.Refresh(outputImage)
 		}
 	}()
 
-	if inputWindow != nil {
-		outputWindow.Show()
-		inputWindow.ShowAndRun()
-		return nil
-	} else {
-		select {}
-	}
+	return state.run()
 }
 
 func add4(bp *byte) {
